+++ test/modules/bundle-widget/index.js
@@ -1,0 +1,11 @@
+module.exports = {
+  extend: '@apostrophecms/widget-type',
+  options: {
+    label: 'Bundle Widget'
+  },
+  webpack: {
+    bundles: {
+      'extra-bundle2': {}
+    }
+  }
+};